main code
This security steward’s software is based on the RSA algorithm to store our various passwords. Only when we enter the correct login password, we can see the passwords corresponding to our different accounts. Otherwise, they are so-called garbled codes. Due to the large prime numbers we use, the probability of hackers breaking the password is reduced. Because it involves information security and privacy, we will continue to maintain the code and test its security.